+++ package.json
@@ -11,10 +11,10 @@
   },
   "dependencies": {
     "date-fns": "^4.1.0",
-    "lodash": "^4.17.21",
     "lucide-react": "^0.525.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-virtuoso": "^4.13.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
