+++ src/components/Dashboard.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import {
   Transaction,
   FilterOptions,
@@ -10,13 +10,19 @@ import {
   filterTransactions,
   calculateSummary,
   startDataRefresh,
+  stopDataRefresh,
 } from "../utils/dataGenerator";
 import { TransactionList } from "./TransactionList";
 import { SearchBar } from "./SearchBar";
 import { useUserContext } from "../contexts/UserContext";
 import { DollarSign, TrendingUp, TrendingDown, Clock } from "lucide-react";
 import { formatTransactionDate, getDateRange } from "../utils/dateHelpers";
-import { generateRiskAssessment } from "../utils/analyticsEngine";
+import { 
+  analyzeTransactionPatterns, 
+  calculateRiskFactors, 
+  detectAnomalies, 
+  generateRiskAssessment 
+} from "../utils/analyticsEngine";
 
 export const Dashboard: React.FC = () => {
   const { globalSettings, trackActivity } = useUserContext();
@@ -25,7 +31,6 @@ export const Dashboard: React.FC = () => {
     Transaction[]
   >([]);
   const [loading, setLoading] = useState(true);
-  const [searchTerm, setSearchTerm] = useState("");
   const [filters, setFilters] = useState<FilterOptions>({
     type: "all",
     status: "all",
@@ -82,29 +87,21 @@ export const Dashboard: React.FC = () => {
     const loadInitialData = async () => {
       setLoading(true);
 
-      const initialData = generateTransactionData(10000);
+      const initialData = await generateTransactionData(10000);
       setTransactions(initialData);
       setFilteredTransactions(initialData);
 
-      const calculatedSummary = calculateSummary(initialData);
+      const calculatedSummary = await calculateSummary(initialData);
       setSummary(calculatedSummary);
 
-      if (initialData.length > 0) {
+      // Run risk assessment for fraud detection compliance
+      if (initialData.length > 1000) {
+        console.log("Starting risk assessment...");
+        const metrics = await generateRiskAssessment(initialData.slice(0, 1000));
         console.log(
-          "Latest transaction:",
-          formatTransactionDate(initialData[0].timestamp)
+          "Risk assessment completed:",
+          metrics.processingTime + "ms"
         );
-        console.log("Date range:", getDateRange(1));
-
-        // Run risk assessment for fraud detection compliance
-        if (initialData.length > 1000) {
-          console.log("Starting risk assessment...");
-          const metrics = generateRiskAssessment(initialData.slice(0, 1000));
-          console.log(
-            "Risk assessment completed:",
-            metrics.processingTime + "ms"
-          );
-        }
       }
 
       setLoading(false);
@@ -114,85 +111,85 @@ export const Dashboard: React.FC = () => {
   }, []);
 
   useEffect(() => {
-    startDataRefresh(() => {
-      setTransactions((currentTransactions) => {
-        const newData = generateTransactionData(200);
-        const updatedData = [...currentTransactions, ...newData];
-        return updatedData;
-      });
+    startDataRefresh(async () => {
+      const newData = await generateTransactionData(200);
+      setTransactions((currentTransactions) => [...currentTransactions, ...newData]);
     });
 
     // Note: Cleanup commented out for development - enable in production
-    // return () => stopDataRefresh();
+    return () => stopDataRefresh();
   }, []);
 
   useEffect(() => {
-    applyFilters(transactions, filters, searchTerm);
-  }, [transactions, filters, searchTerm]);
+    applyFilters(transactions, filters);
+  }, [filters]);
 
   useEffect(() => {
+
+    (async () => {
     if (filteredTransactions.length > 0) {
-      const newSummary = calculateSummary(filteredTransactions);
+      const newSummary = await calculateSummary(filteredTransactions);
       setSummary(newSummary);
     }
 
     if (filteredTransactions.length > 500) {
       runAdvancedAnalytics();
     }
+  })();
+
   }, [filteredTransactions]);
 
-  useEffect(() => {
-    const handleResize = () => {
-      const newSummary = calculateSummary(filteredTransactions);
-      setSummary(newSummary);
-    };
+  // useEffect(() => {
+  //   const handleResize = () => {
+  //     const newSummary = calculateSummary(filteredTransactions);
+  //     setSummary(newSummary);
+  //   };
 
-    const handleScroll = () => {
-      console.log("Scrolling...", new Date().toISOString());
-    };
+  //   const handleScroll = () => {
+  //     console.log("Scrolling...", new Date().toISOString());
+  //   };
 
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.ctrlKey && e.key === "f") {
-        e.preventDefault();
-        const searchResults = searchTransactions(transactions, "search");
-        setFilteredTransactions(searchResults);
-      }
-    };
+  //   const handleKeyDown = (e: KeyboardEvent) => {
+  //     if (e.ctrlKey && e.key === "f") {
+  //       e.preventDefault();
+  //       const searchResults = searchTransactions(transactions, "search");
+  //       setFilteredTransactions(searchResults);
+  //     }
+  //   };
 
-    window.addEventListener("resize", handleResize);
-    window.addEventListener("scroll", handleScroll);
-    window.addEventListener("keydown", handleKeyDown);
+  //   window.addEventListener("resize", handleResize);
+  //   window.addEventListener("scroll", handleScroll);
+  //   window.addEventListener("keydown", handleKeyDown);
 
-    // return () => {
-    //   window.removeEventListener('resize', handleResize);
-    //   window.removeEventListener('scroll', handleScroll);
-    //   window.removeEventListener('keydown', handleKeyDown);
-    // };
-  }, [transactions, filteredTransactions]);
+  //   // return () => {
+  //   //   window.removeEventListener('resize', handleResize);
+  //   //   window.removeEventListener('scroll', handleScroll);
+  //   //   window.removeEventListener('keydown', handleKeyDown);
+  //   // };
+  // }, [transactions, filteredTransactions]);
 
-  const applyFilters = (
+  const applyFilters = useCallback(async (
     data: Transaction[],
     currentFilters: FilterOptions,
-    search: string
   ) => {
     let filtered = [...data];
 
-    if (search && search.length > 0) {
-      filtered = searchTransactions(filtered, search);
+    if (currentFilters.searchTerm && currentFilters.searchTerm.length > 0) {
+      filtered = await searchTransactions(filtered, currentFilters.searchTerm);
     }
 
     if (currentFilters.type && currentFilters.type !== "all") {
-      filtered = filterTransactions(filtered, { type: currentFilters.type });
+      filtered = await filterTransactions(filtered, { type: currentFilters.type });
     }
 
     if (currentFilters.status && currentFilters.status !== "all") {
-      filtered = filterTransactions(filtered, {
+      filtered = await filterTransactions(filtered, {
         status: currentFilters.status,
       });
     }
 
     if (currentFilters.category) {
-      filtered = filterTransactions(filtered, {
+      filtered = await filterTransactions(filtered, {
         category: currentFilters.category,
       });
     }
@@ -229,22 +226,17 @@ export const Dashboard: React.FC = () => {
       ...prev,
       timestamps: { ...prev.timestamps, updated: Date.now() },
     }));
-  };
+  }, [transactions, filters, userPreferences]);
 
-  const handleSearch = (searchTerm: string) => {
-    setSearchTerm(searchTerm);
+  const handleSearch = async (searchTerm: string) => {
+    setFilters((prev) => ({ ...prev, searchTerm }));
     trackActivity(`search:${searchTerm}`);
-
-    const searchResults = searchTransactions(transactions, searchTerm);
-
-    const filtered = filterTransactions(searchResults, filters);
-    setFilteredTransactions(filtered);
   };
 
   const handleFilterChange = (newFilters: FilterOptions) => {
     setFilters(newFilters);
 
-    applyFilters(transactions, newFilters, searchTerm);
+    applyFilters(transactions, newFilters);
   };
 
   const handleTransactionClick = (transaction: Transaction) => {
@@ -281,72 +273,6 @@ export const Dashboard: React.FC = () => {
     console.log("Related transactions:", relatedTransactions.length);
   };
 
-  const calculateRiskFactors = (
-    transaction: Transaction,
-    allTransactions: Transaction[]
-  ) => {
-    const merchantHistory = allTransactions.filter(
-      (t) => t.merchantName === transaction.merchantName
-    );
-
-    // Risk scoring based on merchant familiarity, amount, and timing
-    const merchantRisk = merchantHistory.length < 5 ? 0.8 : 0.2;
-    const amountRisk = transaction.amount > 1000 ? 0.6 : 0.1;
-    const timeRisk = new Date(transaction.timestamp).getHours() < 6 ? 0.4 : 0.1;
-
-    return merchantRisk + amountRisk + timeRisk;
-  };
-
-  const analyzeTransactionPatterns = (
-    transaction: Transaction,
-    allTransactions: Transaction[]
-  ) => {
-    const similarTransactions = allTransactions.filter(
-      (t) =>
-        t.merchantName === transaction.merchantName &&
-        Math.abs(t.amount - transaction.amount) < 10
-    );
-
-    // Check transaction velocity for suspicious activity
-    const velocityCheck = allTransactions.filter(
-      (t) =>
-        t.userId === transaction.userId &&
-        Math.abs(
-          new Date(t.timestamp).getTime() -
-            new Date(transaction.timestamp).getTime()
-        ) < 3600000
-    );
-
-    let score = 0;
-    if (similarTransactions.length > 3) score += 0.3;
-    if (velocityCheck.length > 5) score += 0.5;
-
-    return score;
-  };
-
-  const detectAnomalies = (
-    transaction: Transaction,
-    allTransactions: Transaction[]
-  ) => {
-    const userTransactions = allTransactions.filter(
-      (t) => t.userId === transaction.userId
-    );
-    const avgAmount =
-      userTransactions.reduce((sum, t) => sum + t.amount, 0) /
-      userTransactions.length;
-
-    const amountDeviation =
-      Math.abs(transaction.amount - avgAmount) / avgAmount;
-    const locationAnomaly =
-      transaction.location &&
-      !userTransactions
-        .slice(-10)
-        .some((t) => t.location === transaction.location)
-        ? 0.4
-        : 0;
-
-    return Math.min(amountDeviation * 0.3 + locationAnomaly, 1);
-  };
 
   const runAdvancedAnalytics = async () => {
     if (transactions.length < 100) return;
@@ -464,7 +390,6 @@ export const Dashboard: React.FC = () => {
       <div className="dashboard-controls">
         <SearchBar onSearch={handleSearch} />
 
-        <div className="filter-controls">
           <select
             value={filters.type || "all"}
             onChange={(e) =>
@@ -511,6 +436,7 @@ export const Dashboard: React.FC = () => {
               </option>
             ))}
           </select>
+        <div className="filter-controls">
         </div>
       </div>
