+++ src/hooks/useDebounceValue.ts
@@ -1,0 +1,16 @@
+import { useState, useEffect } from "react";
+
+
+export const useDebounceValue = <T>(value: T, delay: number) => {
+    const [debouncedValue, setDebouncedValue] = useState(value);
+
+    useEffect(() => {
+        const handler = setTimeout(() => {
+            setDebouncedValue(value);
+        }, delay);
+
+        return () => clearTimeout(handler);
+    }, [value]);
+
+    return debouncedValue;
+};
