+++ src/utils/analyticsEngine.ts
@@ -1,7 +1,7 @@
 import { Transaction } from "../types/transaction";
 
 // Comprehensive risk assessment engine for fraud detection and compliance
-export const generateRiskAssessment = (transactions: Transaction[]) => {
+export const generateRiskAssessment = async (transactions: Transaction[]): Promise<any> => {
   const startTime = performance.now();
 
   const fraudScores = transactions.map((transaction) => {
@@ -174,4 +174,73 @@ const analyzeSpendingPattern = (userTransactions: Transaction[]) => {
   });
 
   return { avgAmount, totalAmount, categoryDistribution };
+};
+
+
+
+export const calculateRiskFactors = (
+  transaction: Transaction,
+  allTransactions: Transaction[]
+) => {
+  const merchantHistory = allTransactions.filter(
+    (t) => t.merchantName === transaction.merchantName
+  );
+
+  // Risk scoring based on merchant familiarity, amount, and timing
+  const merchantRisk = merchantHistory.length < 5 ? 0.8 : 0.2;
+  const amountRisk = transaction.amount > 1000 ? 0.6 : 0.1;
+  const timeRisk = new Date(transaction.timestamp).getHours() < 6 ? 0.4 : 0.1;
+
+  return merchantRisk + amountRisk + timeRisk;
+};
+
+export const analyzeTransactionPatterns = (
+  transaction: Transaction,
+  allTransactions: Transaction[]
+) => {
+  const similarTransactions = allTransactions.filter(
+    (t) =>
+      t.merchantName === transaction.merchantName &&
+      Math.abs(t.amount - transaction.amount) < 10
+  );
+
+  // Check transaction velocity for suspicious activity
+  const velocityCheck = allTransactions.filter(
+    (t) =>
+      t.userId === transaction.userId &&
+      Math.abs(
+        new Date(t.timestamp).getTime() -
+          new Date(transaction.timestamp).getTime()
+      ) < 3600000
+  );
+
+  let score = 0;
+  if (similarTransactions.length > 3) score += 0.3;
+  if (velocityCheck.length > 5) score += 0.5;
+
+  return score;
+};
+
+export const detectAnomalies = (
+  transaction: Transaction,
+  allTransactions: Transaction[]
+) => {
+  const userTransactions = allTransactions.filter(
+    (t) => t.userId === transaction.userId
+  );
+  const avgAmount =
+    userTransactions.reduce((sum, t) => sum + t.amount, 0) /
+    userTransactions.length;
+
+  const amountDeviation =
+    Math.abs(transaction.amount - avgAmount) / avgAmount;
+  const locationAnomaly =
+    transaction.location &&
+    !userTransactions
+      .slice(-10)
+      .some((t) => t.location === transaction.location)
+      ? 0.4
+      : 0;
+
+  return Math.min(amountDeviation * 0.3 + locationAnomaly, 1);
 };
